third_party/blink/web_tests/fast/forms/validity-property-expected.txt

Each form control in this document exposes a validity attribute that returns a live instance of ValidityState.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

===> Tests for FIELDSET
PASS "validity" in controls[i] is true
PASS validity1 = controls[i].validity; validity1 instanceof ValidityState is true
PASS validity2 = controls[i].validity; validity1 === validity2 is true
PASS countAttr(controls[i].validity) is 11
===> Tests for SELECT
PASS "validity" in controls[i] is true
PASS validity1 = controls[i].validity; validity1 instanceof ValidityState is true
PASS validity2 = controls[i].validity; validity1 === validity2 is true
PASS countAttr(controls[i].validity) is 11
===> Tests for INPUT
PASS "validity" in controls[i] is true
PASS validity1 = controls[i].validity; validity1 instanceof ValidityState is true
PASS validity2 = controls[i].validity; validity1 === validity2 is true
PASS countAttr(controls[i].validity) is 11
===> Tests for TEXTAREA
PASS "validity" in controls[i] is true
PASS validity1 = controls[i].validity; validity1 instanceof ValidityState is true
PASS validity2 = controls[i].validity; validity1 === validity2 is true
PASS countAttr(controls[i].validity) is 11
===> Tests for BUTTON
PASS "validity" in controls[i] is true
PASS validity1 = controls[i].validity; validity1 instanceof ValidityState is true
PASS validity2 = controls[i].validity; validity1 === validity2 is true
PASS countAttr(controls[i].validity) is 11
===> Tests for OUTPUT
PASS "validity" in controls[i] is true
PASS validity1 = controls[i].validity; validity1 instanceof ValidityState is true
PASS validity2 = controls[i].validity; validity1 === validity2 is true
PASS countAttr(controls[i].validity) is 11
===> Tests for OBJECT
PASS "validity" in controls[i] is true
PASS validity1 = controls[i].validity; validity1 instanceof ValidityState is true
PASS validity2 = controls[i].validity; validity1 === validity2 is true
PASS countAttr(controls[i].validity) is 11
PASS new ValidityState() threw exception TypeError: Failed to construct 'ValidityState': Illegal constructor.
PASS successfullyParsed is true

TEST COMPLETE

    
